+++ kan/__init__.py
@@ -1,1 +1,2 @@
-from .KAN import *
+from .symbolic import *
+from .utils import *
