+++ kan/modeling/spline.py
